refactor(apiv2): move shared request validation
